+++ include/helpers/query.hpp
@@ -10,17 +10,17 @@ using namespace std;
 class Query
 {
 private:
-    size_t number_of_predicates;
+    int64_t number_of_predicates;
 
 public:
     vector<Predicate> predicates;
 
-    Query(vector<float> low, vector<float> high, vector<float> column);
+    Query(vector<float> low, vector<float> high, vector<int64_t> column);
 
     Query(const Query& query);
 
     Query();
 
-    size_t predicate_count();
+    int64_t predicate_count();
 };
 #endif // QUERY_H
