+++ include/helpers/table.hpp
@@ -12,14 +12,14 @@ using namespace std;
 
 class Table{
 private:
-    size_t number_of_rows;
-    size_t number_of_columns;
+    int64_t number_of_rows;
+    int64_t number_of_columns;
 public:
     vector<unique_ptr<Column>> columns;
 
     Table(vector<vector<float> > &columns_to_be_copied);
 
-    Table(size_t number_of_columns);
+    Table(int64_t number_of_columns);
 
     Table(Table *table_to_copy);
 
@@ -27,29 +27,29 @@ public:
 
     ~Table();
 
-    vector<float> materialize_row(size_t row_index);
+    vector<float> materialize_row(int64_t row_index);
 
     void append_column(Column col);
 
     void append(vector<float> row);
 
-    void exchange(size_t index1, size_t index2);
+    void exchange(int64_t index1, int64_t index2);
 
     // Cracks table from position i = low until i == high
     // on column (c) with key (element)
     // Returns in which position the list is greater or equal to key
     // In case of repetition, returns the first one, and all the others come
     // right after it
-    size_t CrackTable(size_t low, size_t high, float element, size_t c);
+    int64_t CrackTable(int64_t low, int64_t high, float element, int64_t c);
 
     // Cracks table in three from position i = low until i == high
     // on column (c) with left key and right key
     // Returns a pair of positions indicating where each patition starts
-    pair<size_t, size_t> CrackTableInThree(size_t low, size_t high, float key_left, float key_right, size_t c);
+    pair<int64_t, int64_t> CrackTableInThree(int64_t low, int64_t high, float key_left, float key_right, int64_t c);
 
-    size_t row_count() const;
+    int64_t row_count() const;
 
-    size_t col_count() const;
+    int64_t col_count() const;
 };
 
 #endif // TABLE_H
