+++ include/indexes/average_kd_tree.hpp
@@ -28,12 +28,12 @@ private:
     int64_t minimum_partition_size = 100;
 
     // Variables to help with initialization
-    vector<size_t> columns;
-    vector<size_t> lower_limits, upper_limits;
+    vector<int64_t> columns;
+    vector<int64_t> lower_limits, upper_limits;
     vector<KDNode*> nodes_to_check;
 
     unique_ptr<KDTree> initialize_index();
 
-    pair<float, size_t> find_average(size_t column, size_t lower_limit, size_t upper_limit);
+    pair<float, int64_t> find_average(int64_t column, int64_t lower_limit, int64_t upper_limit);
 };
 #endif // AVERAGE_KD_TREE_H
