+++ include/indexes/kd_tree/kd_node.hpp
@@ -16,16 +16,16 @@ class KDNode
 public:
 
     float key; // Search-key
-    size_t column; // Search column
+    int64_t column; // Search column
 
     unique_ptr<KDNode> left_child;
     unique_ptr<KDNode> right_child;
 
-    size_t left_position; // Strict smaller than key from this position backwards
-    size_t right_position; // Greater or equal than key from this position forwards
+    int64_t left_position; // Strict smaller than key from this position backwards
+    int64_t right_position; // Greater or equal than key from this position forwards
 
 
-    KDNode(size_t column, float key, size_t left_position, size_t right_position);
+    KDNode(int64_t column, float key, int64_t left_position, int64_t right_position);
     KDNode(const KDNode &node);
     KDNode();
     ~KDNode();
