+++ src/helpers/predicate.cpp
@@ -2,7 +2,7 @@
 #include <cstddef>
 #include "predicate.hpp"
 
-Predicate::Predicate(float low, float high, size_t column)
+Predicate::Predicate(float low, float high, int64_t column)
     : low(low), high(high), column(column){}
 
 Predicate::Predicate() : low(0), high(0), column(0){}
